+++ packages/terra-time-input/src/TimeUtil.js
@@ -269,6 +269,21 @@ class TimeUtil {
 
     return variant;
   }
+
+  static getAriaDescriptionId(options) {
+    const { props, formatDescriptionId, inputAttributes } = options;
+
+    if (props.useExternalFormatMask === false) {
+      if (inputAttributes && inputAttributes['aria-describedby']) {
+        return (`${formatDescriptionId} ${inputAttributes['aria-describedby']}`);
+      }
+      return formatDescriptionId;
+    }
+    if (inputAttributes && inputAttributes['aria-describedby']) {
+      return inputAttributes['aria-describedby'];
+    }
+    return undefined;
+  }
 }
 
 TimeUtil.inputType = {